requirements.txt
streamlit
tensorflow-cpu==2.15.0
numpy
pillow
